Exercise 1: Calculate the Sum of Two Numbers
(Add a and b)

Exercise 2: Check if a Number is Even or Odd

(If entered number is divided by 2
	Print even
else
	Print odd number)

Exercise 3:
Write pseudo code that will perform the following.

a) Read in 5 separate numbers.
b) Calculate the average of the five numbers.
c) Find the smallest (minimum) and largest (maximum) of the five entered numbers.
d) Write out the results found from steps b and c with a message describing what they are

(Your pseudocode here)

-- END --